% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conso_par_groupe.R
\name{conso_par_groupe}
\alias{conso_par_groupe}
\title{Calcul groupe de la consommation}
\usage{
conso_par_groupe(data = conso_region, var_groupe = "secteur")
}
\arguments{
\item{data}{Un data.frame au format de \code{conso_region}.}

\item{var_groupe}{Variable de regroupement (ex : "region", "secteur").}
}
\value{
Un data.frame avec la consommation totale et moyenne par groupe.
}
\description{
Calcule la consommation totale (somme) et la consommation moyenne
(mean) par variable de regroupement.
}
\examples{
conso_par_groupe(var_groupe = "secteur")
}
