% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtrer_conso.R
\name{filtrer_conso}
\alias{filtrer_conso}
\title{Filtrer le jeu de donnees de consommation}
\usage{
filtrer_conso(
  data = conso_region,
  region_choisie = NULL,
  filiere_choisie = NULL
)
}
\arguments{
\item{data}{Un data.frame au format de \code{conso_region}.}

\item{region_choisie}{Nom de la region a conserver (ou NULL).}

\item{filiere_choisie}{Filiere a conserver : "Gaz" ou "Electricite" (ou NULL).}
}
\value{
Un data.frame filtre.
}
\description{
Filtrer le jeu de donnees de consommation
}
\examples{
filtrer_conso(region_choisie = "Bretagne", filiere_choisie = "Gaz")
}
