% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_conso.R
\name{graph_conso}
\alias{graph_conso}
\title{Graphique de la consommation par groupe}
\usage{
graph_conso(
  data = conso_region,
  var_groupe = "secteur",
  titre = "Consommation totale par groupe"
)
}
\arguments{
\item{data}{Un data.frame au format de \code{conso_region}.}

\item{var_groupe}{Variable de regroupement (ex : "region", "secteur").}

\item{titre}{Titre du graphique.}
}
\value{
Un objet \code{ggplot}.
}
\description{
Produit un graphique en barres \code{ggplot2} a partir du resultat de
\code{conso_par_groupe()}.
}
\examples{
graph_conso(var_groupe = "region")
}
